Update dependencies.gradle
Phonograph music player turns the now-playing palette into the most vibrant color of the currently playing song's album art. I am not sure if what I did was the correct way to implement that here, but either way, I hope that you can implement this feature. I am not unsure if phonograph allows users to turn this off, but that would definitely be a good idea. Thank you.